requirements.txt
asgiref==3.11.1
Django==6.0.3
pillow==12.2.0
python-decouple==3.8
sqlparse==0.5.5
